+++ spec/rspec/core/configuration_spec.rb
@@ -1760,18 +1760,21 @@ module RSpec::Core
       end
     end
 
+    def example_numbered(num)
+      instance_double(Example, :id => "./foo_spec.rb[1:#{num}]")
+    end
+
     describe "#force" do
       context "for ordering options" do
-        let(:list) { [1, 2, 3, 4] }
+        let(:list) { 1.upto(4).map { |i| example_numbered(i) } }
         let(:ordering_strategy) { config.ordering_registry.fetch(:global) }
-        let(:rng) { RSpec::Core::RandomNumberGenerator.new config.seed }
-        let(:shuffled) { Ordering::Random.new(config).shuffle list, rng }
+        let(:shuffled) { Ordering::Random.new(config).order list }
 
         specify "CLI `--order defined` takes precedence over `config.order = rand`" do
           config.force :order => "defined"
           config.order = "rand"
 
-          expect(ordering_strategy.order(list)).to eq([1, 2, 3, 4])
+          expect(ordering_strategy.order(list)).to eq(list)
         end
 
         specify "CLI `--order rand:37` takes precedence over `config.order = defined`" do
@@ -1793,7 +1796,7 @@ module RSpec::Core
         specify "CLI `--order defined` takes precedence over `config.register_ordering(:global)`" do
           config.force :order => "defined"
           config.register_ordering(:global, &:reverse)
-          expect(ordering_strategy.order(list)).to eq([1, 2, 3, 4])
+          expect(ordering_strategy.order(list)).to eq(list)
         end
       end
 
@@ -1817,7 +1820,7 @@ module RSpec::Core
 
     describe "#seed_used?" do
       def use_seed_on(registry)
-        registry.fetch(:random).order([1, 2])
+        registry.fetch(:random).order([example_numbered(1), example_numbered(2)])
       end
 
       it 'returns false if neither ordering registry used the seed' do